backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
django-celery-beat==2.6.0
redis==5.0.1
Pillow==10.2.0
python-decouple==3.8
